refactor: namespace integration & unit tests
